Add transaction skeleton
