+++ blockchain/transaction_output.go
@@ -1,0 +1,8 @@
+package blockchain
+
+import "github.com/CAU-CLINK/blockchain_with_go/script"
+
+type TxOutput struct {
+	Value        uint
+	ScriptPubKey script.ScriptPubKey
+}
