+++ blockchain/transaction_output_test.go
@@ -1,0 +1,1 @@
+package blockchain_test
