+++ blockchain/transaction_test.go
@@ -1,0 +1,23 @@
+package blockchain_test
+
+import "testing"
+
+// TODO: Implements test case
+func TestTransaction_IsCoinbase(t *testing.T) {
+
+}
+
+// TODO: Implements test case
+func TestTransaction_Hash(t *testing.T) {
+
+}
+
+// TODO: Implements test case
+func TestNewCoinbase(t *testing.T) {
+
+}
+
+// TODO: Implements test cas
+func TestNewTransaction(t *testing.T) {
+
+}
